:bug: Add empty conftest.py
This commit adds an empty conftest.py file, this is done so that we can
run our tests via the command: `pytest` instead of having to hack it by
running python -m pytest. The reason we had to run `python -m pytest`
before was that otherwise the root folder was not added to the path
by python, which resulted in tests not being able to run. An alternative
solution would have been to add `__init__.py` in the tests dir, however
this is not recommend by pytest when our source code is not a package.